Revert "build: bump deps"
This reverts commit 5d162f4fa46f06ee08c6970d3e08e2e597917360.

+++ tests/common/runner.rs
@@ -77,7 +77,7 @@ pub fn test_execute_alloy<S: Storage + Send + Sync, C: PevmChain + Send + Sync +
         );
 
         assert_eq!(
-            block.header.gas_used as u128,
+            block.header.gas_used,
             tx_results
                 .iter()
                 .last()
